+++ lib/jahuty/snippet.rb
@@ -3,12 +3,12 @@ module Jahuty
     @get
 
     class << self
-      def render(id, params = {})
+      def render(id, options = {})
         raise "API key not set. Did you use Jahuty.key?" unless Jahuty.key?
 
         @get ||= Service::Render.new(Service::Connect.new.call(Jahuty.key))
 
-        @get.call(id, params)
+        @get.call(id, options)
       end
     end
   end