+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change `service/get` to `service/render`.
 - Change `Snippet.get` to `Snippet.render`.
+- Change optional second argument of `Snippet.render` from `params` hash to options hash with `params` key.
 
 ## 1.1.2 - 2020-07-04